refactor(authButtons): make it a component
